Fix stereo audio audio_renderer_ex.dart
Hey,

Mono works, but stereo is broken.
I use this code for my stereo, and it works.

This function is broken. The division by 2 does not fit the stereo bits. The % 4 is checked in the start but accidentally division by 2 is used for the sample amount. Also the left and right sample go to the same destination index by mistake.

NB! I do not use offset, so I do not know whether that is calculated correctly in this one.

Thanks!